Conflicts:
	.gitignore
	modules/cstl-admin/src/main/webapp/views/webservice/edit.html
	modules/storage/provider-data-coveragesgroup/src/main/java/org/constellation/provider/coveragesgroup/util/ConvertersJaxbToGeotk.java
	modules/web-services/map/src/main/java/org/constellation/converter/DataReferenceConverter.java
	modules/web-services/map/src/main/java/org/constellation/map/featureinfo/FeatureInfoUtilities.java
	modules/web-services/map/src/main/java/org/constellation/rest/api/MapServices.java
	modules/web-services/wfs/src/main/java/org/constellation/wfs/ws/DefaultWFSWorker.java
	modules/web-services/wms/src/main/java/org/constellation/map/ws/DefaultWMSWorker.java
	modules/web-services/wmts/src/main/java/org/constellation/wmts/ws/DefaultWMTSWorker.java
	modules/web-services/wps/src/main/java/org/constellation/wps/ws/WPSWorker.java
	modules/xml-bindings/configuration/src/main/java/org/constellation/configuration/utils/WMSPortrayalUtils.java

+++ modules/web-services/csw/internal/src/test/java/org/constellation/metadata/InternalCSWworkerTest.java
@@ -21,12 +21,21 @@
 package org.constellation.metadata;
 
 import java.io.File;
+import java.io.StringWriter;
 import java.util.TimeZone;
 import java.util.logging.Level;
+import java.util.logging.Logger;
+import javax.annotation.PostConstruct;
+import javax.inject.Inject;
+import javax.xml.bind.JAXBElement;
+import javax.xml.bind.Marshaller;
 import javax.xml.bind.Unmarshaller;
 import org.apache.sis.metadata.iso.DefaultMetadata;
 import org.apache.sis.xml.XML;
 import org.constellation.admin.ConfigurationEngine;
+import org.constellation.admin.ProviderBusiness;
+import org.constellation.admin.ServiceBusiness;
+import org.constellation.admin.SpringHelper;
 import org.constellation.admin.dao.ProviderRecord;
 import org.constellation.generic.database.Automatic;
 import org.constellation.provider.DataProviderFactory;
@@ -34,7 +43,7 @@ import org.constellation.provider.DataProviders;
 import static org.constellation.provider.configuration.ProviderParameters.*;
 import static org.constellation.provider.coveragesql.CoverageSQLProviderService.*;
 import org.constellation.test.utils.Order;
-import org.constellation.test.utils.TestRunner;
+import org.constellation.test.utils.SpringTestRunner;
 import org.constellation.util.Util;
 import org.geotoolkit.ebrim.xml.EBRIMMarshallerPool;
 import org.geotoolkit.xml.AnchoredMarshallerPool;
@@ -47,45 +56,59 @@ import org.opengis.parameter.ParameterValueGroup;
  *
  * @author Guilhem Legal (Geomatys)
  */
-@RunWith(TestRunner.class)
+@RunWith(SpringTestRunner.class)
 public class InternalCSWworkerTest extends CSWworkerTest {
 
-    @BeforeClass
-    public static void setUpClass() throws Exception {
-        deleteTemporaryFile();
+    @Inject
+    private ServiceBusiness serviceBusiness;
 
-        pool = EBRIMMarshallerPool.getInstance();
-        fillPoolAnchor((AnchoredMarshallerPool) pool);
+    @Inject
+    private ProviderBusiness providerBusiness;
+    
+    @PostConstruct
+    public void setUpClass() {
+        onlyIso = true;
+        SpringHelper.setApplicationContext(applicationContext);
+        try {
+            if (!serviceBusiness.getServiceIdentifiers("csw").contains("default")) {
+                deleteTemporaryFile();
 
-        final File configDir = ConfigurationEngine.setupTestEnvironement("InternalCSWWorkerTest");
+                pool = EBRIMMarshallerPool.getInstance();
+                fillPoolAnchor((AnchoredMarshallerPool) pool);
 
-        File CSWDirectory  = new File(configDir, "CSW");
-        CSWDirectory.mkdir();
-        final File instDirectory = new File(CSWDirectory, "default");
-        instDirectory.mkdir();
+                final File configDir = ConfigurationEngine.setupTestEnvironement("InternalCSWWorkerTest");
 
-        //we write the data files
-        writeProvider("meta1.xml",         "42292_5p_19900609195600");
-        writeProvider("meta2.xml",         "42292_9s_19900610041000");
-        writeProvider("meta3.xml",         "39727_22_19750113062500");
-        writeProvider("meta4.xml",         "11325_158_19640418141800");
-        writeProvider("meta5.xml",         "40510_145_19930221211500");
-        writeProvider("meta-19119.xml",    "mdweb_2_catalog_CSW Data Catalog_profile_inspire_core_service_4");
-        writeProvider("imageMetadata.xml", "gov.noaa.nodc.ncddc. MODXXYYYYJJJ.L3_Mosaic_NOAA_GMX or MODXXYYYYJJJHHMMSS.L3_NOAA_GMX");
-        writeProvider("ebrim1.xml",        "000068C3-3B49-C671-89CF-10A39BB1B652");
-        writeProvider("ebrim2.xml",        "urn:uuid:3e195454-42e8-11dd-8329-00e08157d076");
-        writeProvider("ebrim3.xml",        "urn:motiive:csw-ebrim");
-        writeProvider("meta13.xml",        "urn:uuid:1ef30a8b-876d-4828-9246-dcbbyyiioo");
+                File CSWDirectory  = new File(configDir, "CSW");
+                CSWDirectory.mkdir();
+                final File instDirectory = new File(CSWDirectory, "default");
+                instDirectory.mkdir();
 
-        //we write the configuration file
-        final String nulll = null;
-        Automatic configuration = new Automatic("internal", nulll);
-        configuration.putParameter("shiroAccessible", "false");
+                //we write the data files
+                writeProvider("meta1.xml",         "42292_5p_19900609195600");
+                writeProvider("meta2.xml",         "42292_9s_19900610041000");
+                writeProvider("meta3.xml",         "39727_22_19750113062500");
+                writeProvider("meta4.xml",         "11325_158_19640418141800");
+                writeProvider("meta5.xml",         "40510_145_19930221211500");
+                writeProvider("meta-19119.xml",    "mdweb_2_catalog_CSW Data Catalog_profile_inspire_core_service_4");
+                writeProvider("imageMetadata.xml", "gov.noaa.nodc.ncddc. MODXXYYYYJJJ.L3_Mosaic_NOAA_GMX or MODXXYYYYJJJHHMMSS.L3_NOAA_GMX");
+                /*writeProvider("ebrim1.xml",        "000068C3-3B49-C671-89CF-10A39BB1B652");
+                writeProvider("ebrim2.xml",        "urn:uuid:3e195454-42e8-11dd-8329-00e08157d076");
+                writeProvider("ebrim3.xml",        "urn:motiive:csw-ebrim");
+                writeProvider("meta13.xml",        "urn:uuid:1ef30a8b-876d-4828-9246-dcbbyyiioo");*/
 
-        ConfigurationEngine.storeConfiguration("CSW", "default", configuration);
+                //we write the configuration file
+                final String nulll = null;
+                Automatic configuration = new Automatic("internal", nulll);
+                configuration.putParameter("shiroAccessible", "false");
 
-        worker = new CSWworker("default");
-        worker.setLogLevel(Level.FINER);
+                serviceBusiness.create("csw", "default", configuration, null, null);
+
+                worker = new CSWworker("default");
+                worker.setLogLevel(Level.FINER);
+            }
+        } catch (Exception ex) {
+            Logger.getLogger(InternalCSWworkerTest.class.getName()).log(Level.SEVERE, null, ex);
+        }
 
     }
 
@@ -208,7 +231,7 @@ public class InternalCSWworkerTest extends CSWworkerTest {
         super.DescribeRecordTest();
     }
 
-    public static void writeProvider(String resourceName, String identifier) throws Exception {
+    public void writeProvider(String resourceName, String identifier) throws Exception {
 
         final DataProviderFactory service = DataProviders.getInstance().getFactory("coverage-sql");
         final ParameterValueGroup source = service.getProviderDescriptor().createValue();
@@ -225,11 +248,22 @@ public class InternalCSWworkerTest extends CSWworkerTest {
 
         Unmarshaller u = pool.acquireUnmarshaller();
         u.setProperty(XML.TIMEZONE, TimeZone.getTimeZone("GMT+2:00"));
-        DefaultMetadata meta = (DefaultMetadata) u.unmarshal(Util.getResourceAsStream("org/constellation/xml/metadata/" + resourceName));
+        Object obj = u.unmarshal(Util.getResourceAsStream("org/constellation/xml/metadata/" + resourceName));
+        if (obj instanceof JAXBElement) {
+            obj = ((JAXBElement)obj).getValue();
+        }
+        DefaultMetadata meta = (DefaultMetadata) obj; 
         pool.recycle(u);
 
-        ConfigurationEngine.writeProvider(identifier, null, ProviderRecord.ProviderType.LAYER, service.getName(), source);
-        ConfigurationEngine.saveProviderMetadata(meta, identifier);
-
+        Marshaller m = pool.acquireMarshaller();
+        m.setProperty(XML.TIMEZONE, TimeZone.getTimeZone("GMT+2:00"));
+        final StringWriter sw = new StringWriter();
+        m.marshal(meta, sw);
+        pool.recycle(m);
+        
+        
+        providerBusiness.createProvider(identifier, null, ProviderRecord.ProviderType.LAYER, service.getName(), source);
+        providerBusiness.updateMetadata(identifier, null, meta.getFileIdentifier(), sw.toString());
+        
     }
 }
